requirements.txt
Flask>=2.0
face_recognition
numpy
Pillow
opencv-python
ImageHash
